backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# File handling
openpyxl==3.1.2
reportlab==4.0.9
python-markdown==3.5.2
aiofiles==23.2.1

# Web scraping
beautifulsoup4==4.12.3
selenium==4.17.2
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# YAML support
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pillow==10.2.0
